#61: Update pages to TypeScript

+++ practice-two/src/pages/Home/Home.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useId, useState } from 'react'
+import React, { useCallback, useId, useState } from 'react'
 import useSWR from 'swr'
 import { toast } from 'react-toastify'
 import { Button, Input, LoadingSpinner, Option, Popup, Product, SearchBar, Select, Tag } from '@components'
@@ -7,6 +7,7 @@ import { productService } from '@services'
 import { useDebounce } from '@hooks'
 import { toastConfig } from '@utils'
 import { MESSAGES } from '@constants'
+import { TProduct } from '@types'
 import styles from './Home.module.css'
 
 const API_PRODUCTS = import.meta.env.VITE_API_PRODUCTS
@@ -23,40 +24,47 @@ const Home = () => {
         return API_SORT_PRODUCTS + sortType
     }
   }, [sortType])
-  const { data: products, error, isLoading, mutate } = useSWR(getProductsEndPoint, productService.get)
+  const { data: products, error, isLoading, mutate } = useSWR<TProduct[], Error>(getProductsEndPoint, productService.get)
   const [searchValue, setSearchValue] = useState('')
   const debouncedValue = useDebounce(searchValue, 1000)
-  const { data: find, isLoading: isFinding } = useSWR(
+  const { data: find, isLoading: isFinding } = useSWR<TProduct[], Error>(
     debouncedValue.length ? [API_PRODUCTS, debouncedValue] : null,
-    ([path, query]) => productService.find(path, query)
+    ([path, query]: string[]) => productService.find(path, query)
   )
 
   const notifyId = useId()
   const [showDeleteDialog, setShowDeleteDialog] = useState(false)
   const [showAddDialog, setShowAddDialog] = useState(false)
   const [showEditDialog, setShowEditDialog] = useState(false)
-  const [selectedProductId, setSelectedProductId] = useState(undefined)
+  const [selectedProductId, setSelectedProductId] = useState<number>()
   const [shouldFetch, setShouldFetch] = useState(false)
-  const { data, isLoading: isFetch } = useSWR(
+  const { data, isLoading: isFetch } = useSWR<TProduct[], Error>(
     shouldFetch ? API_PRODUCTS + `/${selectedProductId}` : null,
     productService.get,
     { revalidateOnFocus: false }
   )
 
   const handleNotifyError = useCallback(
-    (error) => {
-      toast.error(error.message, toastConfig(notifyId, toast.POSITION.TOP_CENTER))
+    (error: Error | string) => {
+      let messages = ''
+      if (error instanceof Error) {
+        messages = error.message
+      } else {
+        messages = error
+      }
+
+      toast.error(messages, toastConfig(notifyId, toast.POSITION.TOP_CENTER))
     },
     [notifyId]
   )
 
-  const handleDeleteProduct = useCallback((id) => {
+  const handleDeleteProduct = useCallback((id: number) => {
     setShowDeleteDialog(true)
     setSelectedProductId(id)
   }, [])
 
   const handleEditProduct = useCallback(
-    (id) => {
+    (id: number) => {
       setSelectedProductId(id)
       setShouldFetch(true)
       if (!isFetch) setShowEditDialog(true)
@@ -73,7 +81,10 @@ const Home = () => {
   const handleCancel = useCallback(() => setShowDeleteDialog(false), [])
 
   const handleConfirm = useCallback(async () => {
-    const remainProducts = products.filter((item) => item.id !== selectedProductId)
+    if (selectedProductId === undefined) return
+    if (!products) return
+
+    const remainProducts = (products).filter((item) => item.id !== selectedProductId)
     const status = await productService.remove(selectedProductId)
 
     if (status !== 200) return handleNotifyError(MESSAGES.DELETE_FAILED)
@@ -86,7 +97,7 @@ const Home = () => {
   }, [products, handleNotifyError, notifyId, mutate, selectedProductId])
 
   const renderProducts = useCallback(
-    (data) => {
+    (data: TProduct[] | null | undefined) => {
       if (!data?.length) return <h1 className={styles.empty}>There no products yet</h1>
 
       return data?.map(({ id, name, description }) => (
@@ -103,9 +114,9 @@ const Home = () => {
     [handleDeleteProduct, handleEditProduct]
   )
 
-  const handleChangeSearchValue = useCallback((e) => setSearchValue(e.target.value), [])
+  const handleChangeSearchValue = useCallback((e: React.ChangeEvent<HTMLInputElement>) => setSearchValue(e.target.value), [])
 
-  const handleSortChange = useCallback((e) => setSortType(e.target.value), [])
+  const handleSortChange = useCallback((e: React.ChangeEvent<HTMLSelectElement>) => setSortType(e.target.value), [])
 
   if (error) return handleNotifyError(error)
